Add react router, refactor BookForm component
Добавлен react router, переработан компонент BookForm, он был разбит на более мелкие компоненты.
