--- src/components/ListView/components/index.js
@@ -1,5 +1,0 @@
-import {BookTemplate} from "./BookTemplate";
-import {BookForm} from "./BookForm";
-import {Header} from "./Header";
-
-export {BookForm, BookTemplate, Header}
