+++ src/pages/BookForm/components/DateContainer/DateContainer.module.scss
@@ -1,0 +1,49 @@
+div.date{
+  width: 80%;
+  gap: 10%;
+
+  display: inline-flex;
+  align-items: flex-end;
+  justify-content: space-between;
+
+  label.label{
+    display: inline-flex;
+    flex-direction: column;
+    gap: 1vh;
+
+    font-size: 2.5vh;
+    width: 45%;
+  }
+
+  input.input{
+    @include inputTemplate();
+  }
+}
+
+
+@media screen and (max-width: 1000px){
+  div.date{
+    width: 80%;
+    gap: 2%;
+
+    label.label{
+      width: 50%;
+    }
+  }
+}
+
+
+@media screen and (max-width: 650px){
+  label.label{
+    font-size: 2vh;
+  }
+
+  div.date{
+    width: 80%;
+    gap: 2%;
+
+    label.label{
+      width: 48%;
+    }
+  }
+}
