+++ src/pages/BookForm/components/DateContainer/index.js
@@ -1,0 +1,1 @@
+export {default as DateContainer} from './DateContainer';
