+++ src/pages/BookForm/components/FormButtons/FormButtons.module.scss
@@ -1,0 +1,39 @@
+div.buttonContainer{
+  width: 80%;
+
+  display: inline-flex;
+  justify-content: space-around;
+
+  button.add, button.delete{
+    width: 30%;
+    height: 5vh;
+
+    cursor: pointer;
+    outline: none;
+
+    background: $gunmetal;
+    border: 2px $ghostWhite solid;
+
+    font-size: 2.5vh;
+    color: $ghostWhite;
+
+    border-radius: 5px;
+    transition: .2s;
+  }
+
+  button.add:hover{
+    border-color: $zomp;
+    color: $zomp;
+  }
+
+  button.delete:hover{
+    border-color: $oldRose;
+    color: $oldRose;
+  }
+}
+
+@media screen and (max-width: 650px){
+  div.buttonContainer > button.add, div.buttonContainer > button.delete{
+    width: 40%;
+  }
+}
