+++ src/pages/BookForm/components/FormButtons/index.js
@@ -1,0 +1,1 @@
+export {default as FormButtons} from './FormButtons';
