+++ src/pages/BookForm/components/index.js
@@ -1,0 +1,5 @@
+import {CloseButton} from "./CloseButton";
+import {DateContainer} from "./DateContainer";
+import {FormButtons} from "./FormButtons";
+
+export {CloseButton, DateContainer, FormButtons};
