+++ src/pages/ListView/ListView.module.scss
@@ -43,7 +43,7 @@ div.buttonContainer{
   box-sizing: border-box;
   padding-inline: calc(3vh + 1%);
 
-  button.button{
+  button.addButton{
     position: relative;
 
     width: 100%;
