+++ src/style.scss
@@ -39,6 +39,25 @@ body{
   justify-content: center;
 }
 
+@mixin inputTemplate(){
+  outline: none;
+  box-sizing: border-box;
+
+  padding-block: 1vh;
+  padding-inline: 2vh;
+
+  font-size: 2.3vh;
+  color: $ghostWhite;
+
+  background: $gunmetal;
+  border: 2px $ghostWhite solid;
+
+  &::placeholder {
+    color: $ghostWhite;
+    opacity: .5;
+  }
+}
+
 @media screen and (max-width: 800px){
   body{
     ::-webkit-scrollbar{
